QRCODE
Added an abstract class for QRcode, which is extended by three other
classes which include the scoreQRcode which has a score, geolocation,
and a comment; LoginQRcode used for logging the player into thier
account; profileQR code for viewing other players profiles.

Qrcode is able to generate both the login and profile qr codes on screen

ScanQRActivity is able to scan QR codes and score them, still needs to
prompt user for geolocation and if they want to add an image or comment.

Geolocation class contains longitude and latitude for mapping of QR
codes